Consolidate shared deref/UUID helpers in helpers.go
derefStr/derefBool/toUUIDs lived in vm_api.go while helpers.go (which
held the other shared helpers) only pointed at them by comment. Move them
into helpers.go next to derefFloat32. Pure relocation, no behavior change.
